charge changed to zeff, more descriptive, more clear.  Currently only works for hydrogen.  initial states seem to work, awaiting testing

+++ findbasis/single_active_electron.hpp
@@ -16,8 +16,8 @@ struct sae_param {
 template <typename T>
 struct sae {
     std::vector < sae_param<T> > params;
-    int Z;
-    int N;
+    double Z;
+    double N;
     T gs_energy;
 };
